requirements.txt
django==4.0
djangorestframework
djangorestframework-simplejwt